feat: created confirmation email page
When a user creates an account, they are now directed to a dedicated confirmation email page.

//TODO:
- Fix the redirection after email is confirmed to the login page
- Add signup error when an account already exists
- Add a generate new code on the confirmation page
